// aimrx-reference/features/basic-emr/types.ts
export type AllergySeverity = DrizzleAllergySeverity;
// ...
export enum AllergySeverityEnum {
  Mild = "mild",
  Moderate = "moderate",
  Severe = "severe",
}
// ...
export interface Allergy {
  readonly id: string;
  patientId?: string;
  encounterId?: string | null;
  name: string;
  reactionType: string;
  severity: AllergySeverity;
  readonly createdAt: Date;
  readonly updatedAt: Date;
}